+++ projects/nst-test/src/app/file-info/file-info.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { GithubFile } from '../file/file.component';
-import { TraverserSelectors, TraversingState } from 'ngx-state-traverser';
+import { TraverserSelectors, TraverserState } from 'ngx-state-traverser';
 
 @Component({
   selector: 'app-file-info',
@@ -11,6 +11,6 @@ export class FileInfoComponent {
 
   context = TraverserSelectors.TraverserContext<GithubFile>(this.store);
 
-  constructor(private readonly store: Store<TraversingState>) { }
+  constructor(private readonly store: Store<TraverserState>) { }
 
 }