+++ projects/nst-test/src/app/folder/folder.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { select, Store } from '@ngrx/store';
-import { TraverserSelectors, TraversingState, TraverserActions } from 'ngx-state-traverser';
+import { TraverserSelectors, TraverserState, TraverserActions } from 'ngx-state-traverser';
 
 @Component({
     selector: 'app-folder',
@@ -10,7 +10,7 @@ export class FolderComponent {
 
     context = this.store.pipe(select(TraverserSelectors.getContext));
 
-    constructor(private readonly store: Store<TraversingState>) { }
+    constructor(private readonly store: Store<TraverserState>) { }
 
     showDetails(path: string) {
         this.store.dispatch(new TraverserActions.LoadTile({tile: 'details', path}));